backend/pyproject.toml
[project]
name = "llama-index-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo",email = "ricky_olalquiaga@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-index-core (>=0.14.19,<0.15.0)",
    "llama-index-llms-openai (>=0.7.0,<0.8.0)",
    "llama-index-llms-openai-like (>=0.7.0,<0.8.0)",
    "llama-index-embeddings-openai (>=0.6.0,<0.7.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "llama-index-workflows (>=2.17.3,<3.0.0)"
]

[project.optional-dependencies]
google = [
    "llama-index-llms-google-genai",
    "llama-index-embeddings-google-genai"
]
anthropic = [
    "llama-index-llms-anthropic (>=0.11.0,<0.12.0)",
]
ollama = [
    "llama-index-llms-ollama (>=0.10.0,<0.11.0)",
]
local-embeddings = [
    "llama-index-embeddings-huggingface (>=0.7.0,<0.8.0)",
]
all = [
    "llama-index-llms-google-genai",
    "llama-index-embeddings-google-genai",
    "llama-index-llms-anthropic (>=0.11.0,<0.12.0)",
    "llama-index-llms-ollama (>=0.10.0,<0.11.0)",
    "llama-index-embeddings-huggingface (>=0.7.0,<0.8.0)",
]

[tool.poetry]
packages = [
    {include = "api", from = "src"},
    {include = "utils", from = "src"},
    {include = "workflow", from = "src"},    
    {include = "guardrails", from = "src"},    
    {include = "integrations", from = "src"},    
    ]

# GRUPOS - Solo para desarrollo (no van en la distribución)
[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"